添加类别接口

+++ live-api/src/main/java/com/me2me/live/service/LiveExtService.java
@@ -24,5 +24,5 @@ public interface LiveExtService {
 	 * @param page 分页
 	 * @return
 	 */
-	Response kingdomByCategory(int uid,int kcid,int page);
+	Response kingdomByCategory(long uid,int kcid,int page);
 }
